% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nc.R
\name{cf_close}
\alias{cf_close}
\title{Close a nc file}
\usage{
cf_close(nc)
}
\arguments{
\item{nc}{An object of class ncdf4 (as returned from nc_open), indicating what file to read from.}
}
\description{
Close a nc file
}

