% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{cf_create}
\alias{cf_create}
\title{Creates a new netCDF file on disk, given the variables the new file is to contain.}
\usage{
cf_create(filename, vars, ...)
}
\arguments{
\item{filename}{Name of the netCDF file to be created.}

\item{vars}{Either an object of class ncvar4 describing the variable to be created, or a vector (or list) of such objects to be created.}

\item{...}{Other arguments pass to nc_create}
}
\description{
The "name_lbl" varibles and "coordinates" attribures will be created according the type of dimension.
The global attribure "Conventions" will be created.
}
