% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nc.R
\name{cf_open}
\alias{cf_open}
\title{Opens an existing netCDF file for reading (or, optionally, writing).}
\usage{
cf_open(filename, write = FALSE)
}
\arguments{
\item{filename}{Name of the existing netCDF file to be opened.}

\item{If}{FALSE (default), then the file is opened read-only. If TRUE, then writing to the file is allowed.}
}
\description{
Opens an existing netCDF file for reading (or, optionally, writing).
}

