% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nc.R
\name{cfarr_nc}
\alias{cfarr_nc}
\title{Convert an array into a new netCDF file on disk, given the variables the new file is to contain.}
\usage{
cfarr_nc(..., filename)
}
\arguments{
\item{...}{variables write into netCDF file}

\item{filename}{Name of the netCDF file to be created.}
}
\description{
An array will be convert into a new netCDF with CF convention
}

