% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{cfdim_def}
\alias{cfdim_def}
\title{Defines a netCDF dimension}
\usage{
cfdim_def(name, units, vals, ...)
}
\arguments{
\item{name}{Name of the dimension to be created (character string).}

\item{units}{The dimension's units (character string).}

\item{vals}{The dimension's values (vector of numeric type or string).
If the type is string, a new dimension "str_len" and variable "name_ldl" 
will be created. The "coordinates" attribures will be added to relativly variables}

\item{...}{Other arguments pass to ncdim_def}
}
\description{
Defines a netCDF dimension
}
