% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nc.R
\name{cfdim_names}
\alias{cfdim_names}
\title{Get all dimension names for a variable}
\usage{
cfdim_names(nc, var)
}
\arguments{
\item{nc}{An object of class ncdf4 (as returned from nc_open), indicating what file to read from.}

\item{var}{The variable name}
}
\description{
Get all dimension names for a variable
}

