% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{cfvar_def}
\alias{cfvar_def}
\title{Defines a netCDF variable}
\usage{
cfvar_def(name, units, dim, ...)
}
\arguments{
\item{name}{Name of the variable  to be created (character string).}

\item{units}{The variable's units (character string).}

\item{dim}{The variable's dimension(s)}

\item{...}{Other arguments pass to ncvar_def}
}
\description{
Defines a netCDF variable
}
