% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nc.R
\name{cfvar_get}
\alias{cfvar_get}
\title{Get varibles values from nc file}
\usage{
cfvar_get(nc, varname, ...)
}
\arguments{
\item{nc}{An object of class ncdf4 (as returned from nc_open), indicating what file to read from.}

\item{varname}{The variables names in the nc file.}

\item{...}{Arguments are used to specify which data will be returned.}
}
\description{
Get varibles values from nc file
}

