% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{cfvar_lables}
\alias{cfvar_lables}
\title{Get labels for all dimensions}
\usage{
cfvar_lables(nc, name, dimension = NULL)
}
\arguments{
\item{nc}{An object of class ncdf4 (as returned from nc_open), indicating what file to read from.}

\item{name}{The variable name}

\item{dimension}{Dimension of input variables}
}
\description{
Get labels for all dimensions
}
