% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{cfvar_lbl}
\alias{cfvar_lbl}
\title{Get all labels for dimension}
\usage{
cfvar_lbl(nc, name)
}
\arguments{
\item{nc}{An object of class ncdf4 (as returned from nc_open), indicating what file to read from.}

\item{name}{The dimesion name}
}
\description{
Get all labels for dimension
}
