% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nc.R
\name{cfvar_names}
\alias{cfvar_names}
\title{Get all variable names from nc file}
\usage{
cfvar_names(nc)
}
\arguments{
\item{nc}{An object of class ncdf4 (as returned from ncOpen), indicating what file to read from.}
}
\description{
Get all variable names from nc file
}

