% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cfvar_lables}
\alias{cfvar_lables}
\title{Get labels for all dimensions}
\usage{
cfvar_lables(nc, name, dimension = NULL)
}
\arguments{
\item{nc}{An object of class ncdf4 (as returned from nc_open), indicating what file to read from.}

\item{name}{The variable name}
}
\description{
Get labels for all dimensions
}

