% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cfvar_names}
\alias{cfvar_names}
\title{Get all variable names from nc file}
\usage{
cfvar_names(nc)
}
\arguments{
\item{nc}{An object of class ncdf4 (as returned from ncOpen), indicating what file to read from.}
}
\description{
Get all variable names from nc file
}

